GPT/requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.16.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
blis==0.7.9
blobfile==2.0.1
boto3==1.26.81
botocore==1.29.81
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
dataclasses-json==0.5.7
decorator==5.1.1
deeplake==3.2.11
dill==0.3.6
distlib==0.3.6
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
frozenlist==1.3.3
greenlet==2.0.2
hub==3.0.1
huggingface-hub==0.12.1
humbug==0.2.8
idna==3.4
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
langchain==0.0.96
langcodes==3.3.0
llama-index==0.4.15
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.0
nltk==3.8.1
numcodecs==0.11.0
numpy==1.24.2
openai==0.26.5
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathos==0.3.0
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
pox==0.3.2
ppft==1.7.6.6
preshed==3.0.8
prompt-toolkit==3.0.38
pure-eval==0.2.2
pyaml==21.10.1
pycares==4.3.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.5
pyglossary==4.5.0
Pygments==2.14.0
PyJWT==2.6.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
pyvis==0.3.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.46
srsly==2.4.6
stack-data==0.6.2
tenacity==8.1.0
thinc==8.1.8
tiktoken==0.2.0
tokenizers==0.13.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.20.0
wasabi==1.1.1
wcwidth==0.2.6
yarl==1.8.2
